🛠️ :: checkStudentExists user 에서 student 로 이동
