+++ data/src/main/java/team/retum/data/remote/datasource/implementation/UserDataSourceImpl.kt
@@ -33,17 +33,6 @@ class UserDataSourceImpl @Inject constructor(
             )
         }.sendRequest()
 
-    override suspend fun checkStudentExists(
-        gcn: Int,
-        name: String,
-    ) = HttpHandler<Unit>()
-        .httpRequest {
-            userApi.checkStudentExists(
-                gcn = gcn,
-                name = name,
-            )
-        }.sendRequest()
-
     override suspend fun verifyEmail(
         email: String,
         authCode: String,
