+++ data/src/main/java/team/retum/data/repository/StudentRepositoryImpl.kt
@@ -6,12 +6,13 @@ import team.retum.data.remote.response.students.toEntity
 import team.retum.domain.entity.student.StudentInformationEntity
 import team.retum.domain.param.students.EditProfileImageParam
 import team.retum.domain.param.students.ResetPasswordParam
-import team.retum.domain.repository.StudentsRepository
+import team.retum.domain.param.user.CheckStudentExistsParam
+import team.retum.domain.repository.StudentRepository
 import javax.inject.Inject
 
-class StudentsRepositoryImpl @Inject constructor(
+class StudentRepositoryImpl @Inject constructor(
     private val studentsDataSource: StudentsDataSource,
-) : StudentsRepository {
+) : StudentRepository {
     override suspend fun fetchStudentInformation(): StudentInformationEntity =
         studentsDataSource.fetchStudentInformation().toEntity()
 
@@ -25,5 +26,14 @@ class StudentsRepositoryImpl @Inject constructor(
 
     override suspend fun editProfileImage(editProfileImageParam: EditProfileImageParam) {
         studentsDataSource.editProfileImage(editProfileImageRequest = editProfileImageParam.toRequest())
+    }
+
+    override suspend fun checkStudentExists(
+        checkStudentExistsParam: CheckStudentExistsParam,
+    ) {
+        studentsDataSource.checkStudentExists(
+            gcn = checkStudentExistsParam.gcn,
+            name = checkStudentExistsParam.name,
+        )
     }
 }
